Refactor: split output test module out and close reason invariant gap
- Move the `output` test module into `crates/cargo-bdd/src/output/tests.rs`
  (with its snapshots), keeping `output.rs` focused on the formatter and well
  under the 400-line ceiling. The module path is unchanged, so the insta
  snapshot names are unaffected.
- Close the `message == None` half of the reason property: assert the ` - `
  reason marker appears iff reasons are included *and* a message exists, so the
  `append_reason` early-return branch is now exercised rather than asserting
  nothing.
